requirements.txt
python-docx
pypdf
python-pptx
openai
anthropic
google-generativeai
pydantic
python-dotenv
typer
rich
streamlit
tiktoken
watchdog